[MCH] various updates
The major modification consists in reducing the number of histograms
in the pedestals task.
There are now single 2D histograms that contain the pedestals and noise
values for all the channels in the readout, one column for each DS board.
The horizontal axis is organized in terms of FEEID/CRU_LINK/DS_NUMBER, while
the vertical axis has one bin for each of the 64 channels of one DS board.